Move payara-nexus-staging repository
Moves the definition of this repository to the "default" repository block and has it enabled/disabled via properties. This keeps the repository order intact by having them polled after all of the standard set of repositories - when they are added by profile they get queried first.

Also adjusts the order of the repositories back in line with Payara 6