FISH-10151 Fix test for Windows
This property now resolves, but on Windows the environment variable is USERNAME.
If the env variable isn't found the validation is delegated to later and null is returned, which evaluates to valid.
This change makes it so if on Windows we check for the correct env variable.
The test is expected to fail on both OS as in all likelihood your username isn't just numbers

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>